requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-generativeai
pytest
httpx
python-multipart
pillow
pre-commit
google-genai
mcp[cli]
langgraph
langchain-mcp-adapters
langchain-google-genai
firebase-admin
